VehiclesRestController -> getVehiclesById
Manager classes

+++ src/main/java/com/doranco/yari/user/ICustomerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ICustomerRepository extends JpaRepository<Customer, Integer> {
+public interface ICustomerRepository extends JpaRepository<Customer, Long> {
 
 
 }
